requirements.txt

crawl4ai>=0.4.0

# Scraping — Fallback
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
playwright>=1.43.0

# LLM & Embeddings
openai>=1.12.0
sentence-transformers>=2.7.0

# Vector search
faiss-cpu>=1.7.4
numpy>=1.26.4

# API
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.4

# Scraping & Content Extraction
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
trafilatura>=1.8.1

# Frontend
streamlit>=1.33.0
watchdog>=4.0.0

# Utilities & Evaluation
pandas>=2.2.2
scikit-learn>=1.4.2
python-dotenv>=1.0.1
tqdm>=4.66.2
